6-size.c prints the sizes of various types in c

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	long long llint;
 	float f;
 
-	printf("Size of a char: %c byte(s)\n", sizeof(c));
+	printf("Size of a char: %c byte(s)\n", (unsigned long)sizeof(c));
 	printf("Size of an int: %i byte(s)\n", sizeof(i));
 	printf("Size of a long int: %li byte(s)\n", sizeof(lint));
-	printf("Size of a long long int: %lli byte(s)\n", sizeof(llint));
-	printf("Size of a float: %f byte(s)\n", sizeof(f));
+	printf("Size of a long long int: %lli byte(s)\n", (unsigned long)sizeof(llint));
+	printf("Size of a float: %f byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
